requirements.txt
python-telegram-bot[webhooks]~=20.0

python-dotenv~=0.21.0
requests~=2.28.1
pytz~=2022.7
PyAutoGUI~=0.9.53

httplib2~=0.21.0
oauth2client~=4.1.3
google-api-python-client~=2.70.0

pymongo~=4.3.3